anv: shader pool size bump for 4GiB pools
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/43926>

+++ src/intel/vulkan/anv_allocator.c
@@ -346,8 +346,8 @@ anv_block_pool_init(struct anv_block_pool *pool,
                     struct anv_device *device,
                     const char *name,
                     uint64_t start_address,
-                    uint32_t initial_size,
-                    uint32_t max_size)
+                    uint64_t initial_size,
+                    uint64_t max_size)
 {
    VkResult result;
